requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements-dev.txt requirements-dev.in
#
autoflake==0.6.6
autopep8==1.2.2
coverage==4.0.3           # via pytest-cov
decorator==4.0.10         # via ipython, traitlets
django-extensions==1.7.5
factory-boy==2.7.0        # via pytest-factoryboy
fake-factory==0.7.2       # via factory-boy
flake8==2.5.4
inflection==0.3.1         # via pytest-factoryboy
ipython-genutils==0.1.0   # via traitlets
ipython==5.1.0
isort==4.2.5
mccabe==0.4.0             # via flake8
pep8==1.7.0               # via autopep8, flake8
pexpect==4.2.1            # via ipython
pickleshare==0.7.4        # via ipython
prompt-toolkit==1.0.9     # via ipython
ptyprocess==0.5.1         # via pexpect
py==1.4.31                # via pytest
pydocstyle==1.0.0
pyflakes==1.0.0           # via autoflake, flake8
pygments==2.1.3           # via ipython
pytest-cov==2.2.1
pytest-django==2.9.1
pytest-factoryboy==1.3.0
pytest==2.9.1
python-dateutil==2.6.0    # via fake-factory
simplegeneric==0.8.1      # via ipython
six==1.10.0               # via django-extensions, fake-factory, prompt-toolkit, python-dateutil, traitlets
traitlets==4.3.1          # via ipython
wcwidth==0.1.7            # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via ipython
